LayoutTests/fast/dom/toString_attributes-expected.txt

Test that the toString() stringifier has the right attributes

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


* HTMLAnchorElement
PASS descriptor.value is an instance of Function
PASS descriptor.writable is true
PASS descriptor.enumerable is true
PASS descriptor.configurable is true

* HTMLAreaElement
PASS descriptor.value is an instance of Function
PASS descriptor.writable is true
PASS descriptor.enumerable is true
PASS descriptor.configurable is true
PASS successfullyParsed is true

TEST COMPLETE
